refactor part 3, almost complete make modules
